pvr: Fix depth repilcate value for load op shaders
`depth_clear_to_reg` specifies the mrt resource index which contains
info on which output register to use rather than the shared reg from
which to load the value from. This was leading to cases where no
shared regs were allocated and no values would be DMAed into the
shareds by the PDS while the USC expected the sh reg for the depth
clear value to be available.

Signed-off-by: Karmjit Mahil <Karmjit.Mahil@imgtec.com>