pvr: Refactor subpass ds and sample count setup
Now we first check the sample count from the ds attachment as well
as setting it up.

For the iteration through the color attachments, a hint is
provided to the compiler with the exit on
`j >= PVR_MAX_COLOR_ATTACHMENTS` to allow for the code to be
unwrapped if the compiler decides to.

Signed-off-by: Karmjit Mahil <Karmjit.Mahil@imgtec.com>