pvr: fix allocation size of clear colour consts shared regs buffer
The number of const shared registers was being used for the allocation size
rather than the number of bytes. In practice this doesn't make a difference as
the max allocation size is 24 bytes, which then gets rounded up to 64 bytes by
the buffer allocation function. However, we might as well make the allocation
size correct to avoid any future confusion. Noticed through code inspection.

Signed-off-by: Frank Binns <frank.binns@imgtec.com>